+++ Sources/SwiftyNifti/Voxel.swift
@@ -2,13 +2,13 @@ import Foundation
 
 public struct Voxel: Equatable {
   
-  public let value: Double
+  public let value: Float
   
-  init(value: Double) {
+  init(value: Float) {
     self.value = value
   }
   
   init(_ uint8: UInt8) {
-    self.value = Double(uint8)
+    self.value = Float(uint8)
   }
 }
